Switch to Int64 indexing.

+++ tensile-dimensions/lib/src/Numeric/Tensile/Dimensions/Dims/Types.hs
@@ -40,6 +40,7 @@
 module Numeric.Tensile.Dimensions.Dims.Types where
 
 import Data.Functor.Identity
+import Data.Int (Int64)
 import Data.Proxy
 import Unsafe.Coerce (unsafeCoerce)
 
@@ -56,7 +57,7 @@ import Numeric.Tensile.Dimensions.Types
 type Dims (ds :: [Nat]) = TypedList Dim ds
 
 instance Eq (Dims ds) where
-    (==) = unsafeCoerce ((==) :: [Int] -> [Int] -> Bool)
+    (==) = unsafeCoerce ((==) :: [Int64] -> [Int64] -> Bool)
     {-# INLINE (==) #-}
 
 instance Show (Dims ds) where
@@ -64,8 +65,8 @@ instance Show (Dims ds) where
     showsPrec p ds = showParen (p >= 10)
       $ showString "Dims " . showsPrec p (fromDims ds)
 
--- | Similar to `natVal` from `GHC.TypeLits`, but returns `Int`.
-fromDims :: Dims ds -> [Int]
+-- | Similar to `natVal` from `GHC.TypeLits`, but returns `Int64`.
+fromDims :: Dims ds -> [Int64]
 fromDims ds = elimDims ds fromDim --Numeric.Tensile.Dimensions.Types.map fromDim
 {-# INLINE fromDims #-}
 
@@ -75,7 +76,7 @@ elimDims (d :+ ds) f = f d : elimDims ds f
 
 -- | Product of all dimension sizes. 
 -- Caution: numerical overflow will lead to undefined behavior.
-size :: Dims ds -> Int
+size :: Dims ds -> Int64
 size = product . fromDims
 {-# INLINE size #-}
 
@@ -150,7 +151,7 @@ reifyDims d k = unsafeCoerce (WithKnownDims k :: WithKnownDims d r) d
 
 newtype WithKnownDims ds r = WithKnownDims (KnownDims ds => r)
 
-unsafeReifyDims :: [Int] -> (forall ds. KnownDims ds => Dims ds -> r) -> r
+unsafeReifyDims :: [Int64] -> (forall ds. KnownDims ds => Dims ds -> r) -> r
 unsafeReifyDims []     f = f S
 unsafeReifyDims (x:xs) f = 
   unsafeReifyDim x $ \p ->
